refactor: move prim factory
