+++ ide-lsp/src/main/java/org/aya/lsp/server/AyaLanguageServer.java
@@ -22,6 +22,7 @@ import org.aya.cli.library.source.MutableLibraryOwner;
 import org.aya.cli.single.CompilerFlags;
 import org.aya.generic.Constants;
 import org.aya.generic.util.AyaFiles;
+import org.aya.ide.LspPrimFactory;
 import org.aya.ide.action.*;
 import org.aya.lsp.actions.LensMaker;
 import org.aya.lsp.actions.SemanticHighlight;
@@ -29,7 +30,6 @@ import org.aya.lsp.actions.SymbolMaker;
 import org.aya.lsp.library.WsLibrary;
 import org.aya.lsp.models.ComputeTermResult;
 import org.aya.lsp.models.HighlightResult;
-import org.aya.lsp.prim.LspPrimFactory;
 import org.aya.lsp.utils.Log;
 import org.aya.lsp.utils.LspRange;
 import org.aya.prettier.AyaPrettierOptions;
